tracker/requirements.txt
pandas
PyGithub>=1.57
pytest>=6.0.0
tabulate>=0.9.0
